Tutor Availibility or smth fixed fully and works as expected

+++ MathBridgeSystem.Application/DTOs/TutorSchedule/SearchScheduleTutorsRequest.cs
@@ -4,13 +4,13 @@ namespace MathBridgeSystem.Application.DTOs.TutorSchedule
 {
     public class SearchAvailableTutorsRequest
     {
-        public int? DaysOfWeek { get; set; }
+        public byte? DaysOfWeek { get; set; }
         public TimeOnly? StartTime { get; set; }
         public TimeOnly? EndTime { get; set; }
         public bool? CanTeachOnline { get; set; }
         public bool? CanTeachOffline { get; set; }
         public DateTime? EffectiveDate { get; set; }
-                public int Page { get; set; } = 1;
+        public int Page { get; set; } = 1;
         public int PageSize { get; set; } = 20;
     }
 }
